move "postgres" module to "psycopg2_connector"
