+++ procrastinate/__init__.py
@@ -1,6 +1,6 @@
 from procrastinate import metadata as _metadata_module
 from procrastinate.app import App
-from procrastinate.postgres import PostgresJobStore
+from procrastinate.psycopg2_connector import PostgresJobStore
 from procrastinate.retry import BaseRetryStrategy, RetryStrategy
 from procrastinate.store import BaseJobStore
