+++ tests/conftest.py
@@ -8,7 +8,7 @@ from psycopg2 import sql
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 
 from procrastinate import app as app_module
-from procrastinate import jobs, migration, postgres, testing, aiopg_connector
+from procrastinate import jobs, migration, psycopg2_connector, testing, aiopg_connector
 
 # Just ensuring the tests are not polluted by environment
 for key in os.environ:
@@ -36,7 +36,9 @@ def setup_db():
             _execute(cursor, "CREATE DATABASE {}", "procrastinate_test_template")
 
     migrator = migration.Migrator(
-        job_store=postgres.PostgresJobStore(dbname="procrastinate_test_template")
+        job_store=psycopg2_connector.PostgresJobStore(
+            dbname="procrastinate_test_template"
+        )
     )
     connection = migrator.job_store.connection
     with closing(connection):
@@ -77,7 +79,7 @@ def connection(connection_params):
 
 @pytest.fixture
 def pg_job_store(connection_params):
-    job_store = postgres.PostgresJobStore(**connection_params)
+    job_store = psycopg2_connector.PostgresJobStore(**connection_params)
     yield job_store
     job_store.connection.close()
